+++ extensions/chroma/lightID.ts
@@ -1173,6 +1173,7 @@ export const LightIDList: {
    LatticeEnvironment: {},
    DaftPunkEnvironment: {},
    HipHopEnvironment: {},
+   ColliderEnvironment: {},
    GlassDesertEnvironment: {
       '0': [1, 2, 3, 4, 5, 6],
       '1': [1, 2, 3, 4, 5, 6, 7, 8],