+++ extensions/chroma/lightIDgenerate.ts
@@ -169,6 +169,7 @@ export const LightIDList: {
    LatticeEnvironment: {},
    DaftPunkEnvironment: {},
    HipHopEnvironment: {},
+   ColliderEnvironment: {},
    GlassDesertEnvironment: {
       0: Array.from(Array(6), (_, i) => i + 1),
       1: Array.from(Array(8), (_, i) => i + 1),